Port to gdext
Godot 4 uses GdExtension instead of GdNative.  The Rust library gdext is
made by the same maintainers of the gdnative crate.

One difference that was kind of important was that the position of a
tile is now a Vector2i instead of a Vector2.  This lead me to change the
coordinate scalar type to a i32.  u32 was dumb to begin with and it
makes for less work converting between my and Godot's types.